requirements.txt
alabaster==0.7.7
Babel==2.2.0
cffi==1.6.0
coverage==4.0.3
cryptography==0.9
docutils==0.12
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
enum34==1.0.4
GeoIP==1.3.2
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.7
Jinja2==2.8
MarkupSafe==0.23
ndg-httpsclient==0.4.0
netaddr==0.7.15
nose==1.3.6
py==1.4.27
pyasn1==0.1.7
pycparser==2.13
pycrypto==2.6.1
pyflakes==0.8.1
Pygments==2.1.3
pyOpenSSL==0.15.1
pytest==2.7.1
pytest-cov==2.2.1
python-dateutil==2.5.1
pytz==2016.1
readline==6.2.4.1
requests==2.7.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
urllib3==1.14
wheel==0.24.0
